kasoft-hub/backend/requirements.txt
fastapi==0.109.0
uvicorn==0.27.0
pyodbc==5.0.1
python-dotenv==1.0.0
apscheduler==3.10.4
pydantic==2.5.3
pydantic-settings==2.1.0
